// frontend/src/api/model/measurements.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { File } from './file';
import type { OrganizationSummary } from './organizationSummary';
import type { UserSummary } from './userSummary';
import type { Flight } from './flight';
import type { ProcessingStatusEnum } from './processingStatusEnum';
import type { MeasurementTypeEnum } from './measurementTypeEnum';

export interface Measurements {
  readonly id: string;
  readonly files: readonly File[];
  readonly owner: OrganizationSummary;
  readonly author: UserSummary;
  readonly flight: Flight;
  is_deleted?: boolean;
  /** @nullable */
  deleted_at?: string | null;
  /** Status of async background processing

* `pending` - Pending processing
* `processing` - Processing in progress
* `completed` - Processing completed
* `failed` - Processing failed */
  processing_status?: ProcessingStatusEnum;
  /** @nullable */
  time_start?: string | null;
  /** @nullable */
  time_end?: string | null;
  readonly time_created: string;
  /** @maxLength 150 */
  name: string;
  description?: string;
  public?: boolean;
  /** Type of measurement

* `D` - Debug measurement
* `S` - Static measurement
* `M` - Mobile measurement (ground)
* `C` - Civil airborne measurement
* `A` - Special airborne measurement */
  measurement_type?: MeasurementTypeEnum;
  /** @nullable */
  base_location_lat?: number | null;
  /** @nullable */
  base_location_lon?: number | null;
  /** @nullable */
  base_location_alt?: number | null;
  /** @nullable */
  deleted_by?: number | null;
  /** Measurement campaigns this measurement belongs to. */
  campaigns?: string[];
}